TelegramBot/requirements.txt
CherryPy==18.9.0
paho_mqtt==2.1.0
Requests==2.32.5
telepot==12.7
